Refactor applications to exclude UserDetailsServiceAutoConfiguration for security improvements
- Updated main application classes in pos-mcp-server, pos-order, pos-people, pos-price, pos-security-service, pos-shop-manager, pos-tax, and pos-vehicle-fitment to exclude UserDetailsServiceAutoConfiguration.
- Enhanced NltiRequestServiceImpl by simplifying span attribute setting and ensuring non-null checks.
- Refactored SelfRegistrationServiceImpl to use constants for error codes and improved code readability.
- Updated WorkorderStateMachine and WorkSessionServiceImpl to use Math.toIntExact for safer numeric conversions.
- Added notes on security assessment findings, including CVE and CWE issues.
- Updated pom.xml to replace spring-security-core with spring-boot-starter-security for better dependency management.

+++ pos-workorder/src/main/java/com/positivity/workorder/internal/service/WorkSessionServiceImpl.java
@@ -134,7 +134,7 @@ public class WorkSessionServiceImpl implements WorkSessionService {
                 .filter(b -> b.getBreakEndAt() != null)
                 .mapToLong(b -> Duration.between(b.getBreakStartAt(), b.getBreakEndAt()).toSeconds())
                 .sum();
-        session.setTotalDurationSeconds((int) Math.max(0, totalSeconds - breakSeconds));
+        session.setTotalDurationSeconds(Math.toIntExact(Math.max(0L, totalSeconds - breakSeconds)));
 
         session = workSessionRepository.save(session);
